fix: security hardening - SSRF protection, auth gates, manifest validation, domain fixes
