+++ package.json
@@ -2,6 +2,7 @@
   "name": "worldwideview",
   "version": "0.1.0",
   "private": true,
+  "workspaces": ["packages/*"],
   "scripts": {
     "clean": "node -e \"const fs=require('fs');fs.rmSync('.next',{recursive:true,force:true})\"",
     "dev": "next dev --webpack",
@@ -12,6 +13,13 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@worldwideview/wwv-plugin-sdk": "*",
+    "@worldwideview/wwv-plugin-aviation": "*",
+    "@worldwideview/wwv-plugin-maritime": "*",
+    "@worldwideview/wwv-plugin-wildfire": "*",
+    "@worldwideview/wwv-plugin-borders": "*",
+    "@worldwideview/wwv-plugin-camera": "*",
+    "@worldwideview/wwv-plugin-military": "*",
     "@auth/prisma-adapter": "^2.11.1",
     "@prisma/adapter-better-sqlite3": "^7.5.0",
     "@prisma/client": "^7.5.0",
