+++ tsconfig.json
@@ -19,7 +19,14 @@
       }
     ],
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@worldwideview/wwv-plugin-sdk": ["./packages/wwv-plugin-sdk/src"],
+      "@worldwideview/wwv-plugin-aviation": ["./packages/wwv-plugin-aviation/src"],
+      "@worldwideview/wwv-plugin-maritime": ["./packages/wwv-plugin-maritime/src"],
+      "@worldwideview/wwv-plugin-wildfire": ["./packages/wwv-plugin-wildfire/src"],
+      "@worldwideview/wwv-plugin-borders": ["./packages/wwv-plugin-borders/src"],
+      "@worldwideview/wwv-plugin-camera": ["./packages/wwv-plugin-camera/src"],
+      "@worldwideview/wwv-plugin-military": ["./packages/wwv-plugin-military/src"]
     }
   },
   "include": [
